Move mutations to queries
